Mostly, this is a refactoring commit.  There is one significant new feature, though: overrides now only work within a class heirarchy, which is to say that a subclass can override an element in a base class, but a child scope cannot otherwise override an element in a base scope.
I've also done a good bit of refactoring, though; notably, AST#evaluate now
takes named arguments, and I changed the 'name' parameter to 'type' in all of
the Component classes (this was all internal, but was confusing as it was).

I also removed the need for the autonaming stuff -- it's now acceptable for
components not to have names, and everything behaves correctly.  I haven't yet
removed the autoname code, though; I'll do that on the next commit.


git-svn-id: https://reductivelabs.com/svn/puppet/trunk@952 980ebf18-57e1-0310-9a29-db15c13687c0