Making some of the metaprogramming a bit more explicit and a bit easier to manage.  In the process, I have created multiple Util modules, only one of which is used for this current commit.
The main work in this commit is that I created a single, common method for dynamically creating classes and moved all of the Class.new users to using this class.  It handles a lot of the complexity for me, and most of the users have just a couple of lines of code, now, instead of tens.

git-svn-id: https://reductivelabs.com/svn/puppet/trunk@1473 980ebf18-57e1-0310-9a29-db15c13687c0