This commit adds two important features (but which probably were not worth the priority I suddenly placed on them).
First, it adds search paths as I originally requested in #114.  There is
now a 'lib' setting, which can be used to tell Puppet where to find
manifests.  Any file you tell Puppet to parse will have its directory
automatically added to the lib path.  Also, Puppet will check the
PUPPETLIB environment variable for further directories to search.

Second, it converts the 'import' mechanism into a normal function, which
means that you can now use variables and what-have-you in it.  Of
course, this function uses the lib mechanism.  This is something that's
always bothered me about the language, and having it fixed means you can
do simple things like have custom code in the top scope for each
operating system and then do "import os/$operatingsystem" to evaluate
that code.  Without this, you would either need a huge case statement or
the code would need to be in a class, which often isn't sufficient.



git-svn-id: https://reductivelabs.com/svn/puppet/trunk@1605 980ebf18-57e1-0310-9a29-db15c13687c0