Fixing #292 (A bug in tagmail that causes any tag other than 'all' to fail) and #277 (tagmail report missing To: header).
#292 was weird because the messages just didn't have the tags at all.
The problem was that states didn't have tags, yet states were the source
of nearly all messages.  So, I added tags to the states, and included
the state name in the tag list.  Also, types were not including the type
name in the tag list, so I added that.  And, of course, a few unit tests
to check it all.



git-svn-id: https://reductivelabs.com/svn/puppet/trunk@1638 980ebf18-57e1-0310-9a29-db15c13687c0