Trying to get the functionality I had in previous tests.  Mostly I want to be able to load many files but run just a couple of methods, which test/unit supports via -n, selectively enabling -d.  I now can do this with the Rakefile, using 'TESTOPTS="-n <method> -d" rake <blah>, although I find that a bit kludgy (certainly more so than 'rake -n <blah> -d <blah>').
However, I also want to be able to automatically determine which test
suites conflict (meaning something in one suite causes a failure in
another), which too-often happens.  So, I'm going to mess around some
with 'test' until I get that, and then see if I can move that
functionality to the rakefile.



git-svn-id: https://reductivelabs.com/svn/puppet/trunk@2146 980ebf18-57e1-0310-9a29-db15c13687c0