Taking a first stab at moving configuration compiling into the indirection system.  There are still quite a few unanswered questions, the two most notable being embodied in unimplemented tests in the Configuration Code terminus.
This also requires changing the behaviour in a few places.
In particular, 'puppet' and the 'module_puppet' cfengine
module need to store a Node object in memory with the appropriate
classes, since that's now the only way to communicate with
the compiler.  That integration work has not yet been done,
partially because the old configuration handler (which the
soon-to-be-deprecated master handler now uses) still exists.