This commit is focused on getting the 'puppet' executable to work.  As a result, it involves a lot of integration-level testing, and a lot of small design changes to make the code actually work.
In particular, indirections can now have default termini,
so that configurations and facts default to their code terminus

Also, I've removed the ability to manually control whether
ast nodes are used.  I might need to add it back in later,
but if so it will be in the form of a global setting,
rather than the previous system of passing it through 10 different
classes.  Instead, the parser detects whether there are AST nodes
defined and requires them if so or ignores them if not.

About 75 tests are still failing in the main set of tests,
but it's going to be a long slog to get them working --
there are significant design issues around them, as most of
the failures are a result of tests trying to emulate both the
client and server sides of a connection, which normally would
have different fact termini but in this case must have the same
terminus just because they're in the same process and are global.

The next step, then, is to figure that process out, thus finding a way
to make this all work.