All tests should now pass again.
This is the first real pass towards using caching.  The `puppet`
executable actually uses the indirection work, instead of
handlers and such (and man! is it cleaner).

Most of this work was a result of trying to get the client-side
story working, with correct yaml caching of configurations, which
means this commit also covers converting configurations to yaml,
which was a much bigger PITA than it needed to be.

I still need to write integration tests, and I also need to cover
the server-side story of a normal configuration retrieval.