I'm working on making file serving work in the indirector now, so I added two abilities to the indirections:  Models can specify a module to extend the indirection instance with, and indirections will use a :select_terminus method, if it's available, to select the terminus to use for finding.  (It's currently only used for finding, not destroying or saving.)
The upshot is that a model can have a module that handles terminus
selection for it, and then extend its indirection with that module.

This will allow me to use the local terminus when the protocol is 'file'
and the REST terminus when the protocol is 'puppet'.  It should also
open the door for other protocols if they become available.