I've now split the file-serving termini into two separate types (in addition to Rest):  A local terminus that just uses direct file paths, and a mounts terminus that uses the file server to figure out what the path should be.
It looks like it also makes sense to split the 'mounts' terminus further,
so there is a 'modules' terminus used to look files up in the terminus.

I've added some integration tests to verify that everything is
hooked together correctly.

Lastly, I added a directory for shared behaviours.  There's a ton of
duplication in this setup, because the Content and Metadata classes
behave almost but not quite identically across the board.