File serving now works.  I've tested a couple of ways to use it, and added integration tests at the most important hook points.
This provides the final class structure for all of these classes,
but a lot of the class names are pretty bad, so I'm planning on
going through all of them (especially the file_server stuff) and
renaming.

The functionality is all here for finding files, though (finally).

Once the classes are renamed, I'll be adding searching ability
(which will enable the recursive file copies) and then adding
the link management and enabling ignoring files.