Intermediate commit.
This commit adds a Request instance into the indirection,
pushing it all the way to the terminus instances.  It's
a big commit because it requires modifying every terminus class.

There are still some thorny design issues.  In particular, who
should be responsible for making the request object?  I've tried
having both the indirection class and the Indirector module creating
it, and both have their issues.

Also, the Catalog class previously allowed passing Node instances
directly to the find method, which is now no longer possible because
the Request class would treat the node as the instance being found.
We need the request class to have two modes, one when it's passed an
instance and one when it's passed a key.