Moving the password file handling into the SSL::Key class.
This was necessary because when the Indirector is used, there
isn't necessarily enough context available to know when a
password file should be used (e.g., when reading a Key from disk,
you don't know if that key was encrypted).

Now, the Key class automatically uses the right password file, and
only tries to use those files that actually exist.

This isn't very flexible, in that it only allows one CA file and
one non-CA file, but no one really uses anything but
the CA file anyway.