Adding configuration support for XMLRPC handlers.
This provides the interface for configuring backward compatibility
in the new-style server code.  It requires a completely different
configuration step, in that REST and xmlrpc handlers are configured
separately, rather than using the same handler names, but this was
considered less evil than having a large abstraction layer for specifying
how the handlers are related.