Fixing the HttpPool module to get rid of an infinite loop.
We can't have the HttpPool class use the Indirector to see
if it has a cert available, because it might be being used to
try to download one, which would cause it to make an http instance,
which would cause it to...   Well, you get the idea.

Adding and fixing a few other tests I ran into on the way.