Always using the cert name to store yaml files, which fixes #1178.
The Master handler previously provided the support for the :node_name
setting, and that functionality has now been moved into the Node
class.  At the same time, the names to search through have been
changed somewhat:  Previously, the certificate name and the
hostname were both used for searching, but now, the cert name
is always searched first (unless node_name == facter), but only
the Facter hostname, domain, and fqdn are used otherwise.  We no
longer split the cert name, only the hostname/domain/fqdn.

In the general case, this provides no behaviour change, because
people's hostname is the same as their certname.  This only
results in a change in behaviour if you specify a certificate
name that is a normal node name, and you want to look that node
up by something other than the full name in the certificate.

Signed-off-by: Luke Kanies <luke@madstop.com>