Adding to the indirection request support for authentication information.
This basically consists of the node name (i.e., the certificate name),
the IP, and whether the request is authenticated (which is determined
by whether it provided a valid certificate).

Now the two REST classes (mongrel and webrick) need to correctly
pass this information to the indirection calls they make.