Removed the unless condition in query, because the issue is a stale cached value and added comments that query will now always do so.
The query method is only called in two places, from inside 'install' in
yum.rb and from inside 'uninstall' in rpm.

This behavior only happens when you have a lower version than the one you
are 'ensuring'.

Since the right package actually gets installed, the system is in sync
next time, but the event of installation is lost and subscribing resources
will never get it.