The REST terminus now uses the content-type and http result codes.
Things are currently broken -- this is a checkpoint commit
so that it's safe to make mistakes and will probably be
removed.

Signed-off-by: Luke Kanies <luke@madstop.com>

The REST terminus now completely does error handling and serialization.

The Integration tests are still completely broken.

Signed-off-by: Luke Kanies <luke@madstop.com>

The Mongrel and Webrick rest handlers no longer yaml-encode exceptions.

They just store the exceptions in plain text in the message body.

They also set the status to 400, rather than 404.

Signed-off-by: Luke Kanies <luke@madstop.com>