Adding rudimentary support for directly managing formats.
We have a simple Format class, and the FormatHandler
module is responsible for managing its instances,
including providing access by mime type or name.

It will soon replace some of the testing functionality
in the FormatHandler module, but for now, it's just
there as a name => mimetype converter, which makes it
seem a lot like overkill.

Signed-off-by: Luke Kanies <luke@madstop.com>