Correctly handling when REST searches return nothing.
We return a 404, which the client can correctly handle.

That wasn't the actual problem causing failing tests, of course --
the problem was that one side of our stub was autodetecting
marshal, and the other side was forcing yaml, which strangely didn't
work.  It was only happening with searching, though, not finding et al.

The 'search returning nil' problem was a bit of a red herring but
needed to be fixed, too.

Signed-off-by: Luke Kanies <luke@madstop.com>