Adding parameter and URL support to the REST terminus.
Previously, the server side correctly pulled parameters out of
the query strings, but the REST terminus never passed them on.  It does
now, at least for finding and searching.  It appears that at least
WEBrick doesn't support parameters for anything other than forms
and GET.

I've also added the ability for the REST terminus to pull
host/port information from the request key, if it's a URI.

Signed-off-by: Luke Kanies <luke@madstop.com>