Refactoring how files in FileServing are named.
Previously, they retained some concept of the URI used to
find them, and this uri was the primary key
for the FileServing instances.  This key was unfortunately
completely useless, as evidenced by the fact that it was never
used except to test that it worked.

I've modified the FileServing instances (through modifying
the Base class) to use their local path as their key, and they
no longer care about the URI at all.

This commit is mostly about fixing the code that interacts with
the instances to use this new API.

Signed-off-by: Luke Kanies <luke@madstop.com>