Adding automatic attribute collection to the new fileserving code.
Basically, this just includes a consistent method for collecting
info (either content or metadata) and then calls that method
when returning instances via the indirector.

It's such a large commit mostly because of small changes in the normal
code and large changes in the testing to accomodate those small changes.

Signed-off-by: Luke Kanies <luke@madstop.com>