lazy load latest package definitions with yumhelper 2.2
merge commit to 0.24x, original commit:

I could observe that yumhelper.py can run quite some time
and use quite a lot of memory (up to 100MB!) if you're
using many repositories. As the yumhelper is only needed
if you use latest as a ensure-keyword it makes no sense
for people like me letting puppet fetch these details
always, if you're never going to use latest.  => lazy loading.

After discussion on the list, this is the new version for
lazy loading yum latest package versions. It is implemented
mainly in the way that Luke proposed. However it stores
the latest informations in the variable latest_info so
the latest method didn't get too hackish and could nearly be
left like it was before.