Fixing and migrating more file tests.
This is an unfortunately messy commit; I should have tried harder
to separate the different tasks into different commits, but it's
not worth going back and doing now.

This is the first commit in the long road of fixing the existing file
tests, and in the process refactoring and better testing the code.

The refactoring in this commit is mostly around the 'stat' instance
variable that tests whether the file exists (and provides its metadata
if it does) and the 'insync?' method in the source property.

Signed-off-by: Luke Kanies <luke@madstop.com>