Refactoring Catalog#add_resource to correctly handle implicit resources.
It now ignores or removes implicit resources that conflict,
and it yields all resources that are valid.

This makes it simple for calling classes to pass in a list
of resources but only perform a chunk of work for valid (i.e.,
non-conflicting) resources.

This refactor is entirely meant as a way of cleaning up the
Transaction#generate interface to the catalog.

Signed-off-by: Luke Kanies <luke@madstop.com>