Fixing a critical bug in the Cacher module.
Basically, the first generated value was always
considered expired the next time it was asked for.
The fix was to create an initial timestamp in the Cacher
module, thus providing a floor for validity.

This is definitely a murky bug, and is especially hard to
test.

Also refactoring the internals just a bit.

Signed-off-by: Luke Kanies <luke@madstop.com>