Refactoring the Cacher interface to always require attribute declaration.
Previously you could dynamically use cached values, but the new interface
requires a single static declaration of the attribute:

    cached_attr(:myattr) { my_init_code() }

This is cleaner, because it makes it easy to turn the code into an init method
and generally makes the whole thing easier to think about.

Most of this commit is going through the different classes that already using the
Caching engine.

Signed-off-by: Luke Kanies <luke@madstop.com>