Add a doc attribute to AST nodes and fill it with the last seen comments
The lexer maintains a stack of last seen comments.
On blank lines the lexer flush the comments.
On each opening brace the lexer enters a new stack level.
On each block AST nodes, the stack is popped.

Each AST nodes has a doc property that is filled with the
last seen comments on node creation (in fact only on important node
creation representing statements).

Signed-off-by: Brice Figureau <brice-puppet@daysofwonder.com>