Supporting a nil expirer on cacher objects.
This works for those classes like Puppet::Type that use an expirer that isn't
always present (they use their catalog as an expirer).

The behaviour is now that if there is no expirer, expire() is a no-op and all
data is considered expired all the time, so it's always fresh.

Signed-off-by: Luke Kanies <luke@madstop.com>