Fix #1560
Adds an rspec test which demonstrates #1560 and a custom 'process'
method for the aliases provider to fix it.

The default processing uses split() to break the line into records on
the separator, which breaks if records can contain the separator. The
custom method I've added uses a 'limited' split() to break the line on
the first separator only.

This commit fixes #1560

Signed-off-by: Paul Lathrop <paul@tertiusfamily.net>