Fixing #1913 - 'undef' resource values do not get copied to the db
There are unfortunately no tests for this patch; the Rails
code is disappointingly low on tests as it is, and it would
have been essentially an herculian effort add all of the necessary
tests just to make sure this worked.

I've verified it works in practice, which should be sufficient for now.

Signed-off-by: Luke Kanies <luke@madstop.com>