Fix #1933 - Inconsistent resource evaluation order in subsequent evaluation runs
While evaluating the AST, catalog vertices are not always ordered
the same way on different run, leading to some tags (which should
have been applied in evaluation order) to not be associated with
some underlying resources.

This changeset change all accesses to resources inside the compiler
to always use an ordered (in evaluation order) list of added
resources.

Signed-off-by: Brice Figureau <brice-puppet@daysofwonder.com>