Fixed #1884 - exported defines are collected by the exporting host
This was caused by the fix to #1472.  That fix unexported
any resources collected from the local catalog.

The crux of this fix is that it separates 'exported'
and 'virtual' a bit more.  It also removes no-longer-needed
functionality where resources copied their virtual or
exported bits from the enclosing define or class.  This is
now obsolete because we don't evaluate virtual defined resources.

The crux of this commit is that defined resources can stay
exported even though they're evaluated, and that exported state
won't inherit to contained resources such that those then don't
get evaluated.

Signed-off-by: Luke Kanies <luke@madstop.com>