Fixing #1543 - Nagios parse errors no longer kill Puppet
The problem wasn't actually transactions, it was how
exceptions were raised in Naginator.  Well, parse errors
actually resulted in an 'exit', rather than an exception,
and the exceptions that Naginator was raising were not
caught by a normal begin block (SyntaxError, rather than
RuntimeError).

This commit raises a RuntimeError-derived error rather
than exiting.

It also adds some context to the error when Puppet catches it.

Signed-off-by: Luke Kanies <luke@madstop.com>