Refixing #1420 - _naginator_name is only used for services
According to:

http://nagios.sourceforge.net/docs/3_0/customobjectvars.html

the special underscore parameters are only supported
on hosts, contacts, and services.  This commit reverts
most of the changes that set Nagios types to use such
a parameter as the namevar.

The original commit should have been broken into two commits:
one to reorganize the file, and the other to make these changes.
As it was, the revert was much harder than it should have been.

Signed-off-by: Luke Kanies <luke@madstop.com>