Fixing #1869 - autoloaded files should never leak exceptions
Ruby's exception hierarchy is a bit strange, in that only
exceptions that sub RuntimeError are caught by default.
This patch explicitly catches the base class, Exception,
which means that LoadError, SyntaxError, and any
RuntimeErrors will all be caught.

This is done for both load() and loadall(); load() uses
Kernel.load, but loadall() uses Kernel.require.

Signed-off-by: Luke Kanies <luke@madstop.com>